+++ tests/test_connector.py
@@ -6,6 +6,7 @@ from tenacity import Retrying
 from sekoia_automation.connector import Connector
 from sekoia_automation.constants import CHUNK_BYTES_MAX_SIZE, EVENT_BYTES_MAX_SIZE
 from sekoia_automation.exceptions import TriggerConfigurationError
+from tests.utils import match_events
 
 EVENTS = ["foo", "bar"]
 
@@ -26,7 +27,9 @@ def test_connector(storage, mocked_trigger_logs):
     test_connector.log = Mock()
     test_connector.log_exception = Mock()
 
-    return test_connector
+    yield test_connector
+
+    test_connector.stop()
 
 
 def test_forward_events(test_connector):
@@ -96,12 +99,23 @@ def test_push_event_to_intake_with_2_events(test_connector, mocked_trigger_logs)
 def test_push_event_to_intake_with_chunks(test_connector, mocked_trigger_logs):
     url = "https://intake.sekoia.io/batch"
     test_connector.configuration.chunk_size = 1
-    mocked_trigger_logs.post(url, json={"event_ids": ["001"]})
-    result = test_connector.push_events_to_intakes(EVENTS)
+    mocked_trigger_logs.post(
+        url, json={"event_ids": ["001"]}, additional_matcher=match_events("foo")
+    )
+    mocked_trigger_logs.post(
+        url, json={"event_ids": ["002"]}, additional_matcher=match_events("bar")
+    )
+    mocked_trigger_logs.post(
+        url, json={"event_ids": ["003"]}, additional_matcher=match_events("baz")
+    )
+    mocked_trigger_logs.post(
+        url, json={"event_ids": ["004"]}, additional_matcher=match_events("oof")
+    )
+    result = test_connector.push_events_to_intakes(["foo", "bar", "baz", "oof"])
     assert result is not None
-    assert len(result) == 2
-    assert mocked_trigger_logs.call_count == 2
-    assert result == ["001", "001"]
+    assert len(result) == 4
+    assert mocked_trigger_logs.call_count == 4
+    assert result == ["001", "002", "003", "004"]
 
 
 def test_push_events_to_intakes_no_events(test_connector):