+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Small fixes in cli commands
+- Better handling of error on data path access
 
 ## [1.0.0] - 2023-02-16